Fixes #1 updating OMNeT++ to 5.4.1
Thanks to @martinpfannemueller for finding a solution.
Other changes include updates to avoid using deprecated OMNeT++ API, and to avoid rounding errors due to new vehavior of par().

+++ src/modules/MTBrownoutServer.h
@@ -50,7 +50,7 @@ class MTBrownoutServer : public MTServer
      * 2500 is the time at which we can visually observe in a real run that the service
      * time converges to the stable value.
      */
-    double cacheRequestCount;
+    int cacheRequestCount;
 
     /**
      * The difference in service time between the start of the run and
